Finished most of the program and fixed things
I have fixed the problem with large folders, I have fixed various things
with the logging and backup form. The backup form now shows a control
box to allow the user to manually close the form, instead after 3
seconds, and the time taken is now logged. Also, I have added a manual
option for the backups.