Added debug mode, current status and time +1 more
The debug mode has been added, this will make it so that the
CurrentInterval logging can be disabled, the current status and time
functions have been added, so you can now see if it has been queued and
you can see the current minute and hour. I have Added a function whereby
the icon in the taskbar will now show a progress bar, unfortunately the
only way I could get this to work was by adding another library.